Navigation Improvement
A significant issue noted was the poor distribution of my navigation elements.

I have started improving everything here.

I added an event listener to make sure the DOMContentLoaded and thereafter added the window change to the code.

I had an issue previously with this but it looks to solve the problem