Axes created
to determine my xMax and yMax I loop through the api data to find the greatest two numbers.

Thereafter I set the xMax and yMax according

This is also displayed on webpage.

I now need to make sure the date format is correct as the api only accepts YYYY/MM/DD and with my code I for a "0" in. If my assignment is marked next month I am going to have an issue with this.